requirements.txt
Flask
Flask-SQLAlchemy
Flask-Login
Flask-Bootstrap5
python-dotenv
requests
Flask-WTF
BeautifulSoup4
nltk
pandas
yfinance
Flask-Paginate
gunicorn
werkzeug
wtforms
pytesseract
python-docx
Pillow
pymysql
pytz